# openapi.yaml
openapi: 3.0.3
info:
  title: Zynox Tracking API
  description: Discord Live Tracking API — track users, bots, guilds, and presence in real time via the Discord Gateway.
  version: 2.0.0
  contact:
    url: https://api.zynoxbot.online
servers:
  - url: https://api.zynoxbot.online/api/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for public endpoints
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for developer dashboard and admin
...
paths:
...
  /stripe/create-checkout:
    post:
      tags: [Billing]
      summary: Create Stripe checkout session
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan, successUrl, cancelUrl]
              properties:
                plan:
                  type: string
                  enum: [pro, enterprise]
                successUrl:
                  type: string
                  format: uri
                cancelUrl:
                  type: string
                  format: uri
      responses:
        "200":
          description: Checkout URL
  /stripe/webhook:
    post:
      tags: [Billing]
      summary: Stripe webhook (server-to-server)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed

  # ─── Admin ───────────────────────────────────────────────
...